pdferator/requirements.txt
PyMuPDF
Pillow
pix2tex[torch]
PyYAML
beautifulsoup4
lxml

